pyproject.toml
[tool.poetry]
name = "tilt-machine"
version = "0.1.0"
description = ""
authors = ["Jacob Nesbitt <jjnesbitt2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PySerial = "^3.4"
python-xlib = "^0.27"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = "^19.10b0"
flake8-black = "^0.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
